Extract junctions from a BAM file

Given a BAM file, extract junction information including co-ordinates, strand, anchor length for each junction read. For details on the format of the output TSV file, check https://github.com/ChristopherWilks/megadepth#junctions.

bam_to_junctions(
  bam_file,
  prefix = file.path(tempdir(), basename(bam_file)),
  all_junctions = TRUE,
  junctions = FALSE,
  long_reads = FALSE,
  filter_in = 65535,
  filter_out = 260,
  overwrite = FALSE
)

Arguments

bam_file

A character(1) with the path to the input BAM file.

prefix

A character(1) specifying the output file prefix. This function creates a file called prefix.jxs.tsv. By default, the prefix is the BAM file name and the file is created in the tempdir() and will be deleted after you close your R session.

all_junctions

A logical(1) indicating whether to obtain all junctions.

junctions

A logical(1) indicating whether to obtain co-occurring jx coordinates.

long_reads

A logical(1) indicating whether to increase the buffer size to accommodate for long-read RNA-sequencing.

filter_in

A integer(1) used to filter in read alignments. See https://github.com/ChristopherWilks/megadepth#bamcram-processing and https://samtools.github.io/hts-specs/SAMv1.pdf for further documentation on how to apply this parameter.

filter_out

A integer(1) used to filter out read alignments. See https://github.com/ChristopherWilks/megadepth#bamcram-processing and https://samtools.github.io/hts-specs/SAMv1.pdf for further documentation on how to apply this parameter.

overwrite

A logical(1) specifying whether to overwrite the output file(s), if they exist already.

Value

A character(1) with the path to the output junction tsv file.

Examples

## Install if necessary
install_megadepth()
#> The latest megadepth version is 1.2.0
#> This is not an interactive session, therefore megadepth has been installed temporarily to 
#> /tmp/RtmprYXJM0/megadepth

## Find the example BAM file
example_bam <- system.file("tests", "test.bam",
    package = "megadepth", mustWork = TRUE
)

## Run bam_to_junctions()
example_jxs <- bam_to_junctions(example_bam, overwrite = TRUE)

## Path to the output file generated by bam_to_junctions()
example_jxs
#>                            all_jxs.tsv 
#> "/tmp/RtmprYXJM0/test.bam.all_jxs.tsv"