Extract junctions from a BAM file

Given a BAM file, extract junction information including co-ordinates, strand, anchor length for each junction read. For details on the format of the output TSV file, check https://github.com/ChristopherWilks/megadepth#junctions.

bam_to_junctions(
  bam_file,
  prefix = file.path(tempdir(), basename(bam_file)),
  all_junctions = TRUE,
  junctions = FALSE,
  long_reads = FALSE,
  overwrite = FALSE
)

Arguments

bam_file
A character(1) with the path to the input BAM file.

prefix
A character(1) specifying the output file prefix. This function creates a file called prefix.jxs.tsv. By default, the prefix is the BAM file name and the file is created in the tempdir() and will be deleted after you close your R session.

all_junctions
A logical(1) indicating whether to obtain all junctions.

junctions
A logical(1) indicating whether to obtain co-occurring jx coordinates.

long_reads
A logical(1) indicating whether to increase the buffer size to accommodate for long-read RNA-sequencing.

overwrite
A logical(1) specifying whether to overwrite the output file(s), if they exist already.

Value

A character(1) with the path to the output junction tsv file.

Examples

## Install if necessary
install_megadepth()

#> The latest megadepth version is 1.1.0b

#> megadepth has been installed to /github/home/bin

## Find the example BAM file
example_bam <- system.file("tests", "test.bam",
    package = "megadepth", mustWork = TRUE
)

## Run bam_to_junctions()
example_jxs <- bam_to_junctions(example_bam, overwrite = TRUE)

## Path to the output file generated by bam_to_junctions()
example_jxs

#>                            all_jxs.tsv 
#> "/tmp/Rtmp99StWT/test.bam.all_jxs.tsv"