Compute coverage summarizations across a set of regions

Given an input set of annotation regions, compute coverage summarizations using Megadepth for a given BigWig file.

get_coverage(
  bigwig_file,
  op = c("sum", "mean", "max", "min"),
  annotation,
  prefix = file.path(tempdir(), "bw.mean")
)

Arguments

bigwig_file
A character(1) with the path to the input BigWig file.

op
A character(1) specifying the summarization operation to perform.

annotation
A character(1) path to a BED file.

prefix
A character(1) specifying the output file prefix. This function creates a file called prefix.annotation.tsv that can be read again later with read_coverage(). By default the file is created in the tempdir() and will be deleted after you close your R session.

Value

A GRanges-class object with the coverage summarization across the annotation ranges.

Details

Note that the chromosome names (seqnames) in the BigWig file and the annotation file should use the same format. Otherwise, Megadepth will return 0 counts.

See also

Examples

## Install if necessary
install_megadepth()

#> The latest megadepth version is 1.0.4

#> megadepth has been installed to /github/home/bin

## Locate example BigWig and annotation files
example_bw <- system.file("tests", "test.bam.all.bw",
    package = "megadepth", mustWork = TRUE
)
annotation_file <- system.file("tests", "testbw2.bed",
    package = "megadepth", mustWork = TRUE
)

## Compute the coverage
bw_cov <- get_coverage(example_bw, op = "mean", annotation = annotation_file)
bw_cov

#> GRanges object with 4 ranges and 1 metadata column:
#>         seqnames          ranges strand |       cov
#>            <Rle>       <IRanges>  <Rle> | <numeric>
#>   [1]      chr10            0-10      * |     0.000
#>   [2]      chr10 8756697-8756762      * |    15.846
#>   [3]      chr10 4359156-4359188      * |     3.000
#>   [4] GL000219.1   168500-168620      * |     1.258
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome; no seqlengths

## If you want to cast this into a RleList object use the following code:
## (it's equivalent to rtracklayer::import.bw(as = "RleList"))
GenomicRanges::coverage(bw_cov, weight = "cov")

#> RleList of length 2
#> $chr10
#> numeric-Rle of length 8756762 with 4 runs
#>   Lengths: 4359155      33 4397508      66
#>   Values :   0.000   3.000   0.000  15.846
#> 
#> $GL000219.1
#> numeric-Rle of length 168620 with 2 runs
#>   Lengths: 168499    121
#>   Values :  0.000  1.258
#> 

## Checking other tools
bed <- rtracklayer::import(annotation_file)
bw_cov_rtrack <- rtracklayer::import.bw(rtracklayer::BigWigFile(example_bw), which = bed)
bw_cov_rtrack

#> GRanges object with 11 ranges and 1 metadata column:
#>          seqnames          ranges strand |     score
#>             <Rle>       <IRanges>  <Rle> | <numeric>
#>    [1]      chr10            1-10      * |         0
#>    [2]      chr10 4359157-4359159      * |         4
#>    [3]      chr10 4359160-4359185      * |         3
#>    [4]      chr10 4359186-4359188      * |         2
#>    [5]      chr10 8756698-8756699      * |         0
#>    [6]      chr10 8756700-8756714      * |         6
#>    [7]      chr10 8756715-8756761      * |        20
#>    [8]      chr10         8756762      * |         0
#>    [9] GL000219.1   168501-168544      * |         0
#>   [10] GL000219.1   168545-168619      * |         2
#>   [11] GL000219.1          168620      * |         1
#>   -------
#>   seqinfo: 2 sequences from an unspecified genome

names(example_bw) <- "example"
fullCov <- derfinder::fullCoverage(rtracklayer::BigWigFileList(example_bw), chrs = GenomeInfoDb::seqlevels(bw_cov))

#> No methods found in package ‘IRanges’ for request: ‘values<-’ when loading ‘derfinder’

#> No methods found in package ‘IRanges’ for request: ‘values’ when loading ‘derfinder’

#> 2020-09-30 16:18:17 fullCoverage: processing chromosome chr10

#> 2020-09-30 16:18:17 loadCoverage: finding chromosome lengths

#> 2020-09-30 16:18:17 loadCoverage: loading BigWig file /usr/local/lib/R/host-site-library/megadepth/tests/test.bam.all.bw

#> 2020-09-30 16:18:17 loadCoverage: applying the cutoff to the merged data

#> 2020-09-30 16:18:17 filterData: originally there were 130694993 rows, now there are 130694993 rows. Meaning that 0 percent was filtered.

#> 2020-09-30 16:18:17 fullCoverage: processing chromosome GL000219.1

#> 2020-09-30 16:18:17 loadCoverage: finding chromosome lengths

#> 2020-09-30 16:18:17 loadCoverage: loading BigWig file /usr/local/lib/R/host-site-library/megadepth/tests/test.bam.all.bw

#> 2020-09-30 16:18:17 loadCoverage: applying the cutoff to the merged data

#> 2020-09-30 16:18:17 filterData: originally there were 179198 rows, now there are 179198 rows. Meaning that 0 percent was filtered.

regionCov <- derfinder::getRegionCoverage(fullCov, regions = bed)

#> 2020-09-30 16:18:17 getRegionCoverage: processing chr10

#> 2020-09-30 16:18:17 getRegionCoverage: done processing chr10

#> 2020-09-30 16:18:18 getRegionCoverage: processing GL000219.1

#> 2020-09-30 16:18:18 getRegionCoverage: done processing GL000219.1

## We have to round the mean to make them comparable
testthat::expect_equivalent(
    round(sapply(regionCov[c(1, 3:4, 2)], function(x) mean(x$value)), 3),
    bw_cov$cov,
)

## Note that there's no need to round here
testthat::expect_equivalent(
    sapply(regionCov[c(1, 3:4, 2)], function(x) sum(x$value)),
    get_coverage(example_bw, op = "sum", annotation = annotation_file)$cov,
)