Read a junction TSV file created by Megadepth as a table

Read an *jxs.tsv file created by bam_to_junctions() or manually by the user using Megadepth.

read_junction_table(tsv_file)

Arguments

tsv_file
A character(1) specifying the path to the tab-separated (TSV) file created manually using megadepth_shell() or on a previous bam_to_junctions() run.

Value

A tibble::tibble() with the junction data that follows the format specified at https://github.com/ChristopherWilks/megadepth#megadepth-pathtobamfile---junctions.

Examples

## Install if necessary
install_megadepth()

#> The latest megadepth version is 1.0.8b

#> megadepth has been installed to /github/home/bin

## Find the example BAM file
example_bam <- system.file("tests", "test.bam",
    package = "megadepth", mustWork = TRUE
)

## Run bam_to_junctions()
example_jxs <- bam_to_junctions(example_bam, overwrite = TRUE)

# ## Read the data as a tibble using the format specified at
# ## https://github.com/ChristopherWilks/megadepth#megadepth-pathtobamfile---junctions
# read_junction_table(example_jxs)